apps/tramMinewSync/requirements.txt
requests
pyodbc
python-dotenv
deep-translator
cryptography